fix int to long

+++ Catebi.Api.Domain/Features/AdoptionBot/Models/AtUser.cs
@@ -9,7 +9,7 @@ public class AtUser
     public int UserId { get; set; }
     public string Name { get; set; }
     public string Telegram { get; set; }
-    public int TelegramChatId { get; set; }
+    public long TelegramChatId { get; set; }
     public bool UsePayedAccount { get; set; }
     public bool IsVolunteer { get; set; }
